services/auto-drama-ai/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
volcengine-python-sdk[ark]
fastapi==0.115.6
uvicorn[standard]==0.34.0

pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==5.0.0

